exp/upbox: create package from exp/cmd/upbox internals
This leaves exp/cmd/upbox a simple wrapper around the upbox package.

The plan is to extend upbox.Schema so that it can be used in the
cmd/upspin tests, and more.
